+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/intellij/runner/functions/component/NewAppServicePlanDialog.java
@@ -26,9 +26,9 @@ import com.intellij.openapi.ui.ValidationInfo;
 import com.intellij.ui.ListCellRendererWrapper;
 import com.microsoft.azure.management.appservice.PricingTier;
 import com.microsoft.azure.management.appservice.SkuName;
-import com.microsoft.azure.management.resources.Location;
-import com.microsoft.azuretools.core.mvp.model.AzureMvpModel;
+import com.microsoft.azure.management.resources.fluentcore.arm.Region;
 import com.microsoft.azuretools.core.mvp.model.function.AzureFunctionMvpModel;
+import com.microsoft.azuretools.core.mvp.model.webapp.AzureWebAppMvpModel;
 import com.microsoft.intellij.ui.components.AzureDialogWrapper;
 import com.microsoft.intellij.util.ValidationUtils;
 import com.microsoft.tooling.msservices.components.DefaultLoader;
@@ -46,13 +46,15 @@ public class NewAppServicePlanDialog extends AzureDialogWrapper {
 
     public static final String CONSUMPTION = "Consumption";
     public static final PricingTier CONSUMPTION_PRICING_TIER = new PricingTier("Consumption", "");
-    private static final String DEFAULT_LOCATION = "eastus";
+    private static final String DEFAULT_REGION = "eastus";
     private static final String RECOMMEND_SUFFIX = " (Recommended)";
+    private static final String MISSING_PRICING_TIER_PROMPT = "Please set pricing tier for app service plan";
+    private static final String MISSING_REGION_PROMPT = "Please set region for app service plan";
 
     private JPanel contentPane;
     private JPanel pnlNewAppServicePlan;
     private JComboBox cbPricing;
-    private JComboBox cbLocation;
+    private JComboBox cbRegion;
     private JTextField txtAppServicePlanName;
 
     private String subscriptionId;
@@ -67,11 +69,11 @@ public class NewAppServicePlanDialog extends AzureDialogWrapper {
 
         this.subscriptionId = subscriptionId;
 
-        cbLocation.setRenderer(new ListCellRendererWrapper() {
+        cbRegion.setRenderer(new ListCellRendererWrapper() {
             @Override
             public void customize(JList list, Object object, int i, boolean b, boolean b1) {
-                if (object instanceof Location) {
-                    setText(((Location) object).name());
+                if (object instanceof Region) {
+                    setText(((Region) object).label());
                 } else if (object instanceof String) {
                     setText(object.toString());
                 }
@@ -91,9 +93,9 @@ public class NewAppServicePlanDialog extends AzureDialogWrapper {
             }
         });
 
-        onLoadPricingTier();
-        onLoadLocation(subscriptionId);
+        cbPricing.addItemListener(e -> onLoadRegion());
 
+        onLoadPricingTier();
         init();
     }
 
@@ -115,6 +117,12 @@ public class NewAppServicePlanDialog extends AzureDialogWrapper {
         } catch (IllegalArgumentException iae) {
             res.add(new ValidationInfo(iae.getMessage(), txtAppServicePlanName));
         }
+        if (cbPricing.getSelectedItem() == null) {
+            res.add(new ValidationInfo(MISSING_PRICING_TIER_PROMPT, cbPricing));
+        }
+        if (cbRegion.getSelectedItem() == null) {
+            res.add(new ValidationInfo(MISSING_REGION_PROMPT, cbRegion));
+        }
         return res;
     }
 
@@ -122,7 +130,7 @@ public class NewAppServicePlanDialog extends AzureDialogWrapper {
     protected void doOKAction() {
         appServicePlan = new AppServicePlanPanel.AppServicePlanWrapper(
                 txtAppServicePlanName.getText(),
-                (Location) cbLocation.getSelectedItem(),
+                (Region) cbRegion.getSelectedItem(),
                 (PricingTier) cbPricing.getSelectedItem());
         super.doOKAction();
     }
@@ -133,23 +141,24 @@ public class NewAppServicePlanDialog extends AzureDialogWrapper {
         super.doCancelAction();
     }
 
-    private void onLoadLocation(String sid) {
-        cbLocation.removeAllItems();
-        Observable.fromCallable(() -> AzureMvpModel.getInstance().listLocationsBySubscriptionId(sid))
+    private void onLoadRegion() {
+        cbRegion.removeAllItems();
+        Observable.fromCallable(() -> AzureWebAppMvpModel
+                .getInstance().getAvailableRegions(subscriptionId, (PricingTier) cbPricing.getSelectedItem()))
                   .subscribeOn(Schedulers.newThread())
-                  .subscribe(locations -> DefaultLoader.getIdeHelper().invokeLater(() -> fillLocation(locations)));
+                  .subscribe(locations -> DefaultLoader.getIdeHelper().invokeLater(() -> fillRegion(locations)));
     }
 
-    private void fillLocation(List<Location> locationList) {
-        if (CollectionUtils.isEmpty(locationList)) {
+    private void fillRegion(List<Region> regionList) {
+        if (CollectionUtils.isEmpty(regionList)) {
             return;
         }
-        locationList.stream().forEach(location -> cbLocation.addItem(location));
-        final Location defaultLocation =
-                locationList.stream()
-                            .filter(location -> StringUtils.equalsAnyIgnoreCase(location.name(), DEFAULT_LOCATION))
-                            .findFirst().orElse(locationList.get(0));
-        cbLocation.setSelectedItem(defaultLocation);
+        regionList.stream().forEach(region -> cbRegion.addItem(region));
+        final Region defaultLocation =
+                regionList.stream()
+                          .filter(region -> StringUtils.equalsAnyIgnoreCase(region.name(), DEFAULT_REGION))
+                          .findFirst().orElse(regionList.get(0));
+        cbRegion.setSelectedItem(defaultLocation);
     }
 
     /**