+++ Utils/hdinsight-node-common/src/com/microsoft/azure/hdinsight/sdk/common/HttpObservable.java
@@ -457,17 +457,19 @@ public class HttpObservable implements ILogger {
                 .map(resp -> this.convertJsonResponseToObject(resp, clazz));
     }
 
-    public Observable<CloseableHttpResponse> executeReqAndCheckStatus(HttpEntityEnclosingRequestBase req, int validStatueCode, List<NameValuePair> pairs) {
-        return request(req, req.getEntity(), pairs, Arrays.asList(getDefaultHeaderGroup().getAllHeaders()))
-                .doOnNext(
-                        resp -> {
-                            int statusCode = resp.getStatusLine().getStatusCode();
-                            if (statusCode != validStatueCode) {
-                                Exceptions.propagate(new UnknownServiceException(
-                                        String.format("Exceute request with unexpected code %s and resp %s", statusCode, resp)
-                                ));
-                            }
-                        }
-                );
+    public Observable<CloseableHttpResponse> executeReqAndCheckStatus(HttpRequestBase httpRequest,
+                                                                      @Nullable HttpEntity entity,
+                                                                      List<NameValuePair> parameters,
+                                                                      List<Header> addOrReplaceHeaders,
+                                                                      int validStatueCode) {
+        return request(httpRequest, entity, parameters, addOrReplaceHeaders)
+                .doOnNext(resp -> {
+                    int statusCode = resp.getStatusLine().getStatusCode();
+                    if (statusCode != validStatueCode) {
+                        Exceptions.propagate(new UnknownServiceException(
+                                String.format("Exceute request with unexpected code %s and resp %s", statusCode, resp)
+                        ));
+                    }
+                });
     }
 }