+++ Utils/hdinsight-node-common/src/com/microsoft/azure/hdinsight/sdk/storage/adlsgen2/ADLSGen2FSOperation.java
@@ -22,26 +22,35 @@
 
 package com.microsoft.azure.hdinsight.sdk.storage.adlsgen2;
 
+import com.google.common.collect.ImmutableList;
 import com.microsoft.azure.hdinsight.sdk.common.HttpObservable;
 import com.microsoft.azure.hdinsight.sdk.rest.azure.storageaccounts.RemoteFile;
 import com.microsoft.azure.hdinsight.sdk.rest.azure.storageaccounts.api.GetRemoteFilesResponse;
 import com.microsoft.azuretools.azurecommons.helpers.NotNull;
+import com.microsoft.azuretools.azurecommons.helpers.Nullable;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.http.Header;
 import org.apache.http.NameValuePair;
 import org.apache.http.client.methods.HttpPatch;
 import org.apache.http.client.methods.HttpPut;
 import org.apache.http.entity.BufferedHttpEntity;
 import org.apache.http.entity.ContentType;
 import org.apache.http.entity.InputStreamEntity;
+import org.apache.http.message.BasicHeader;
 import rx.Observable;
 
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.util.Collections;
 import java.util.List;
 
 public class ADLSGen2FSOperation {
+    public static final String DEFAULT_UMASK = "0000";
+    public static final String PERMISSIONS_HEADER = "x-ms-permissions";
+    public static final String UMASK_HEADER = "x-ms-umask";
+
     private HttpObservable http;
 
     @NotNull
@@ -78,15 +87,43 @@ public class ADLSGen2FSOperation {
                 .setAction("flush");
     }
 
-    public Observable<Boolean> createDir(String dirpath) {
-        HttpPut req = new HttpPut(dirpath);
-        return http.executeReqAndCheckStatus(req, 201, this.createDirReqParams)
-                .map(ignore -> true);
+    public Observable<Boolean> createDir(String dirPath) {
+        return createDir(dirPath, null);
+    }
+
+    public Observable<Boolean> createDir(String dirPath, String permission) {
+        return createDir(dirPath, permission, DEFAULT_UMASK);
+    }
+
+    public Observable<Boolean> createDir(String dirPath, @Nullable String permission, @Nullable String uMask) {
+        HttpPut req = new HttpPut(dirPath);
+        // We will filter out these headers if OAuth is used as authorization method.
+        // Check class ADLSGen2OAuthHttpObservable for more details
+        final List<Header> headers = permission != null && uMask != null
+                                     ? ImmutableList.of(new BasicHeader(PERMISSIONS_HEADER, permission),
+                                                        new BasicHeader(UMASK_HEADER, uMask))
+                                     : Collections.emptyList();
+        return http.executeReqAndCheckStatus(req, null, this.createDirReqParams, headers, 201)
+                   .map(ignore -> true);
     }
 
     public Observable<Boolean> createFile(String filePath) {
+        return createFile(filePath, null);
+    }
+
+    public Observable<Boolean> createFile(String filePath, String permission) {
+        return createFile(filePath, permission, DEFAULT_UMASK);
+    }
+
+    public Observable<Boolean> createFile(String filePath, @Nullable String permission, @Nullable String uMask) {
         HttpPut req = new HttpPut(filePath);
-        return http.executeReqAndCheckStatus(req, 201, this.createFileReqParams)
+        // We will filter out these headers if OAuth is used as authorization method.
+        // Check class ADLSGen2OAuthHttpObservable for more details
+        final List<Header> headers = permission != null && uMask != null
+                                     ? ImmutableList.of(new BasicHeader(PERMISSIONS_HEADER, permission),
+                                                        new BasicHeader(UMASK_HEADER, uMask))
+                                     : Collections.emptyList();
+        return http.executeReqAndCheckStatus(req, null, this.createFileReqParams, headers, 201)
                 .map(ignore -> true);
     }
 
@@ -118,10 +155,9 @@ public class ADLSGen2FSOperation {
             long len = entity.getContentLength();
 
             HttpPatch req = new HttpPatch(filePath);
-            req.setEntity(entity);
             http.setContentType("application/octet-stream");
 
-            return http.executeReqAndCheckStatus(req, 202, this.appendReqParams)
+            return http.executeReqAndCheckStatus(req, entity, this.appendReqParams, Collections.emptyList(), 202)
                     .map(ignore -> len);
         } catch (FileNotFoundException e) {
             throw new RuntimeException(new IllegalArgumentException("Can not find the aritifact"));
@@ -135,7 +171,7 @@ public class ADLSGen2FSOperation {
         List<NameValuePair> flushReqParams = this.flushReqParamsBuilder.setPosition(flushLen).build();
         http.setContentType("application/json");
 
-        return http.executeReqAndCheckStatus(req, 200, flushReqParams)
+        return http.executeReqAndCheckStatus(req, null, flushReqParams, Collections.emptyList(), 200)
                 .map(ignore -> true);
     }
 }